clientes/pyproject.toml
[project]
name = "clientes"
version = "0.1.0"
description = "CRUD de clientes"
authors = [
    {name = "ralapont",email = "rafael.alapont.gomez@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[all] (>=0.128.5,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "motor (>=3.7.1,<4.0.0)",
    "pymongo (>=4.16.0,<5.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
